organise preferences settings UI
